restructured middleend directory
